payment updates

+++ src/components/shared/client/Invoice.tsx
@@ -4,36 +4,52 @@ import {
   useGetEukapayInvoice,
 } from "@/lib/react-query/queries"
 import { Models } from "appwrite"
-import { Dot, ExternalLink, Trash2 } from "lucide-react"
+import { ExternalLink, Trash2 } from "lucide-react"
 import { Link } from "react-router-dom"
 import { toast } from "sonner"
 import Loader from "../Loader"
+import { Badge } from "@/components/ui/badge"
+import { useConfirm } from "../AlertDialogProvider"
 
 const Invoice = ({ invoice }: { invoice: Models.Document }) => {
-  const { data: invoiceData, isPending: isPendingInvoice } =
+  const { data: eukapayInvoice, isPending: isPendingInvoice } =
     useGetEukapayInvoice(invoice.code)
   const { mutateAsync: deleteDocument } = useDeleteDocument()
+  const confirm = useConfirm()
 
   const getStatus = (status: string) => {
     switch (status) {
       case "Paid":
         return (
-          <span className="flex gap-0.5 pr-3 pl-1 py-1.5 text-[0.9rem] border border-border rounded-lg">
-            <Dot className="w-6 h-6 text-green-500 scale-150" />
-            Accepted
-          </span>
+          <Badge
+            variant="outline"
+            className="rounded-md text-[0.65rem] px-1.5 py-0.5 bg-green-100 text-green-600"
+          >
+            {status}
+          </Badge>
         )
       default:
         return (
-          <span className="flex gap-0.5 pr-3 pl-1 py-1.5 text-[0.9rem] border border-border rounded-lg">
-            <Dot className="w-6 h-6 text-amber-400 scale-150" />
+          <Badge
+            variant="outline"
+            className="rounded-md text-[0.65rem] px-1.5 py-0.5 bg-yellow-100 text-yellow-600"
+          >
             {status}
-          </span>
+          </Badge>
         )
     }
   }
 
-  const handleDelete = (documentId: string) => {
+  const handleDelete = async (documentId: string) => {
+    const deleteConfirmed = await confirm({
+      title: `Deleting ${invoice?.title} payment`,
+      body: "Are you sure you want to do that?",
+      cancelButton: "Cancel",
+      actionButton: "Delete",
+    })
+
+    if (!deleteConfirmed) return
+
     try {
       const deletePromise = deleteDocument(documentId)
       toast.promise(deletePromise, {
@@ -50,25 +66,48 @@ const Invoice = ({ invoice }: { invoice: Models.Document }) => {
     return <Loader className="justify-start h-8" />
   }
 
+  const createdAt = new Date(invoice.$createdAt).toLocaleDateString("en-US", {
+    year: "numeric",
+    month: "short",
+    day: "numeric",
+  })
+
   return (
     <div className="flex items-center justify-between gap-6">
-      <div className="flex gap-4">
-        <Link to={invoiceData.paymentUrl} target="_blank">
-          <Button type="button" size="sm">
-            {invoice.title}
-            <ExternalLink className="h-4 w-4 ml-2" />
-          </Button>
-        </Link>
-        {getStatus(invoiceData.status)}
+      <div className="flex w-full justify-between gap-6">
+        <div className="flex flex-col gap-1 mr-4">
+          <div className="flex items-center gap-2">
+            <p className="text-sm font-medium">{invoice.title}</p>
+            {getStatus(eukapayInvoice.status)}
+          </div>
+          <p className="text-xs text-muted-foreground">Created {createdAt}</p>
+        </div>
+        <div className="flex gap-6">
+          {eukapayInvoice.paymentUrl && (
+            <Link to={eukapayInvoice.paymentUrl} target="_blank">
+              <Button type="button" size="sm">
+                EukaPay
+                <ExternalLink className="h-4 w-4 ml-2" />
+              </Button>
+            </Link>
+          )}
+          {invoice.link && (
+            <Link to={invoice.link} target="_blank">
+              <Button type="button" size="sm">
+                Stripe
+                <ExternalLink className="h-4 w-4 ml-2" />
+              </Button>
+            </Link>
+          )}
+        </div>
       </div>
       <div className="flex gap-4">
-        <p className="text-sm">{invoice.status}</p>
         <Button
           type="button"
           variant="outline"
           size="icon"
           onClick={() => handleDelete(invoice.$id)}
-          className="w-14"
+          className="w-10"
         >
           <Trash2 className="h-4 w-4" />
         </Button>
